requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
asgiref==3.8.1
    # via django
asttokens==3.0.0
    # via stack-data
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.2.0
    # via nbconvert
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   pip-tools
    #   typer
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
django==5.1.4
    # via
    #   -r requirements.in
    #   django-tailwind
    #   djangorestframework
django-tailwind==3.8.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
docopt==0.6.2
    # via pipreqs
et-xmlfile==2.0.0
    # via openpyxl
executing==2.1.0
    # via stack-data
fastjsonschema==2.21.1
    # via nbformat
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
ipython==8.12.3
    # via pipreqs
jedi==0.19.2
    # via ipython
jinja2==3.1.5
    # via nbconvert
jsonschema==4.23.0
    # via nbformat
jsonschema-specifications==2024.10.1
    # via jsonschema
jupyter-client==8.6.3
    # via nbclient
jupyter-core==5.7.2
    # via
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.3.0
    # via nbconvert
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mistune==3.1.0
    # via nbconvert
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.4
    # via pipreqs
nbformat==5.10.4
    # via
    #   nbclient
    #   nbconvert
numpy==2.2.1
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==24.2
    # via
    #   build
    #   nbconvert
    #   plotly
pandas==2.2.3
    # via -r requirements.in
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==7.4.1
    # via -r requirements.in
pipreqs==0.5.0
    # via -r requirements.in
platformdirs==4.3.6
    # via jupyter-core
plotly==5.24.1
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via ipython
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psycopg2==2.9.10
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via
    #   ipython
    #   nbconvert
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
pyzmq==26.2.0
    # via jupyter-client
redis==5.2.1
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   yarg
rich==13.9.4
    # via
    #   -r requirements.in
    #   typer
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
tenacity==9.0.0
    # via
    #   -r requirements.in
    #   plotly
tinycss2==1.4.0
    # via nbconvert
tornado==6.4.2
    # via jupyter-client
traitlets==5.14.3
    # via
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typer==0.15.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via typer
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.45.1
    # via pip-tools
yarg==0.1.9
    # via pipreqs
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
